# Config loading optimization