--- a/.gitignore
+++ b/.gitignore
@@ -1,3 +1,4 @@
+.benchmarks
 /.idea
 /temp
 outputs