--- a/news/1328.feature
+++ b/news/1328.feature
@@ -1,0 +1,1 @@
+Improve performance of config composition in a benchmark by 64%