--- a/requirements/dev.txt
+++ b/requirements/dev.txt
@@ -10,6 +10,7 @@ nox
 packaging
 pre-commit
 pytest
+pytest-benchmark
 pytest-snail
 setuptools
 towncrier