--- a/requirements/requirements.txt
+++ b/requirements/requirements.txt
@@ -1,3 +1,3 @@
-omegaconf==2.1.0dev14
+omegaconf>=2.1.0dev16
 antlr4-python3-runtime==4.8
 importlib-resources;python_version<'3.9'