Closes #1328
Closes #1323 (because it's including the fix in omegaconf==2.1.0dev16).

This includes significant optimizations in OmegaConf as well (https://github.com/omry/omegaconf/pull/481, https://github.com/omry/omegaconf/pull/485).

1. Adding a benchmark for config initialization.
2. Config load time reduced by 64% (81ms to 29ms).
3. Example app start time reduced by 28% (308ms to 221ms)
4. Tests time reduced by 28% (122s to 87s).